Get display sort of working

+++ GBEmulator.Core/Models/Object.cs
@@ -2,8 +2,8 @@ namespace GBEmulator.Core.Models;
 
 public class Object
 {
-    public byte YPosition { get; set; }
-    public byte XPosition { get; set; }
+    public int YPosition { get; set; }
+    public int XPosition { get; set; }
     public byte TileIndex { get; set; }
     public byte Attributes { get; set; }
 }
